% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_operations.R
\name{addJob}
\alias{addJob}
\title{Add a job to the specified pool.}
\usage{
addJob(jobId, ...)
}
\arguments{
\item{jobId}{A string that uniquely identifies the job within the account.}

\item{...}{Further named parameters
\itemize{
 \item{"resourceFiles"}: {A list of files that the Batch service will download to the compute node before running the command line.}
 \item{"args"}: {Arguments in the foreach parameters that will be used for the task running.}
 \item{"packages"}: {A list of packages that the Batch service will download to the compute node.}
 \item{"envir"}: {The R environment that the task will run under.}
}}
}
\value{
The request to the Batch service was successful.
}
\description{
Add a job to the specified pool.
}
\examples{
addJob(job-001, packages = c("devtools", "httr"))
}

