% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_operation.R
\name{addTask}
\alias{addTask}
\title{Add a task to the specified job.}
\usage{
addTask(jobId, taskId = "default", ...)
}
\arguments{
\item{jobId}{The id of the job to which the task is to be added.}

\item{taskId}{A string that uniquely identifies the task within the job.}

\item{...}{Further named parameters
\itemize{
 \item{"resourceFiles"}: {A list of files that the Batch service will download to the compute node before running the command line.}
 \item{"args"}: {Arguments in the foreach parameters that will be used for the task running.}
 \item{"packages"}: {A list of packages that the Batch service will download to the compute node.}
 \item{"envir"}: {The R environment that the task will run under.}
}}
}
\value{
The response of task
}
\description{
Add a task to the specified job.
}
\examples{
addTask(job-001, task-001)
}

