% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_operations.R
\name{deleteJob}
\alias{deleteJob}
\title{Deletes a job.}
\usage{
deleteJob(jobId)
}
\arguments{
\item{jobId}{The id of the job to delete..}
}
\value{
The request to the Batch service was successful.
}
\description{
Deletes a job.
}
\details{
Deleting a job also deletes all tasks that are part of that job, and all job statistics. This also overrides the retention period for task data; that is, if the job contains tasks which are still retained on compute nodes, the Batch services deletes those tasks' working directories and all their contents.
}
\examples{
deleteJob(job-001)
}

