% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_operations.R
\name{getJob}
\alias{getJob}
\title{Gets information about the specified job.}
\usage{
getJob(jobId)
}
\arguments{
\item{jobId}{The id of the job.}
}
\value{
A response containing the job.
}
\description{
Gets information about the specified job.
}
\examples{
getJob(job-001)
}

