% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_operations.R
\name{updateJob}
\alias{updateJob}
\title{Updates the properties of the specified job.}
\usage{
updateJob(jobId, ...)
}
\arguments{
\item{jobId}{The id of the job.}
}
\value{
The request to the Batch service was successful.
}
\description{
Updates the properties of the specified job.
}

